+++ macos/Runner/MainFlutterWindow.swift
@@ -23,6 +23,13 @@ class MainFlutterWindow: NSWindow {
                 Task {
                    await transcribe(audioFilePath: audioFilePath, modelPath: modelPath, result: result)
                 }
+            case "detectLanguage":
+                guard let args = call.arguments as? [String: Any] else { return }
+                let audioFilePath = args["audioFilePath"] as! String
+                let modelPath = args["modelPath"] as! String
+                Task {
+                   await detectLanguage(audioFilePath: audioFilePath, modelPath: modelPath, result: result)
+                }
             default:
                 result(FlutterMethodNotImplemented)
             }